fix(related_items): change npc_rows iterator type to u16
Reverts the `as u32` cast in `npc_rows` and updates the return type to `impl Iterator<Item = u16>` to correctly match the underlying data type and resolve clippy warnings while maintaining compilation.

+++ ultros-frontend/ultros-app/src/components/related_items.rs
@@ -304,9 +304,9 @@ fn Recipe(recipe: &'static Recipe, item_id: ItemId) -> impl IntoView {
     }.into_any())
 }
 
-fn npc_rows(npc: &ENpcBase) -> impl Iterator<Item = u32> + '_ {
+fn npc_rows(npc: &ENpcBase) -> impl Iterator<Item = u16> + '_ {
     // TODO- can I just parse the csv into a vec?
-    npc.e_npc_data.iter().map(|i| i.0 as u32)
+    npc.e_npc_data.iter().map(|i| i.0)
 }
 
 fn gil_shop_to_npc(gil_shops: &[GilShopId]) -> Vec<(GilShopId, &'static ENpcBase)> {
